Move states and timezones out of pudl.constants
* Took the enumeration of valid US state & territories, and approximate timezones based
  on state, and relocated them into the pudl.metadata subpackage.
* Removed the data_years dictionary, whose functionality has been superceded by the
  metadata that's stored in the datastore, indicating which years/states of data are
  available for each dataset.
* Removed null strings and whitespace entries from the prime mover code column of the
  generation_fuel_eia923 table, as this was violating the new ENUM constraint.
* Moved the timezone inference function and data into the pudl.transform.eia modules
  since that's the only place that it's used.
