+++ src/pudl/transform/eia.py
@@ -19,15 +19,145 @@ found in :func:`pudl.transform.eia._boiler_generator_assn`.
 
 import importlib.resources
 import logging
+from typing import Dict
 
 import networkx as nx
 import numpy as np
 import pandas as pd
+import timezonefinder
 
 import pudl
-from pudl import constants as pc
+import pudl.constants as pc
 
 logger = logging.getLogger(__name__)
+
+TZ_FINDER = timezonefinder.TimezoneFinder()
+"""A global TimezoneFinder to cache geographies in memory for faster access."""
+
+APPROXIMATE_TIMEZONES: Dict[str, str] = {
+    "AK": "America/Anchorage",      # Alaska
+    "AL": "America/Chicago",        # Alabama
+    "AR": "America/Chicago",        # Arkansas
+    "AS": "Pacific/Pago_Pago",      # American Samoa; Not in CEMS
+    "AZ": "America/Phoenix",        # Arizona
+    "CA": "America/Los_Angeles",    # California
+    "CO": "America/Denver",         # Colorado
+    "CT": "America/New_York",       # Connecticut
+    "DC": "America/New_York",       # District of Columbia
+    "DE": "America/New_York",       # Delaware
+    "FL": "America/New_York",       # Florida (split state)
+    "GA": "America/New_York",       # Georgia
+    "GU": "Pacific/Guam",           # Guam; Not in CEMS
+    "HI": "Pacific/Honolulu",       # Hawaii; Not in CEMS
+    "IA": "America/Chicago",        # Iowa
+    "ID": "America/Denver",         # Idaho (split state)
+    "IL": "America/Chicago",        # Illinois
+    "IN": "America/New_York",       # Indiana (split state)
+    "KS": "America/Chicago",        # Kansas (split state)
+    "KY": "America/New_York",       # Kentucky (split state)
+    "LA": "America/Chicago",        # Louisiana
+    "MA": "America/New_York",       # Massachusetts
+    "MD": "America/New_York",       # Maryland
+    "ME": "America/New_York",       # Maine
+    "MI": "America/Detroit",        # Michigan (split state)
+    "MN": "America/Chicago",        # Minnesota
+    "MO": "America/Chicago",        # Missouri
+    "MP": "Pacific/Guam",           # Northern Mariana Islands; Not in CEMS
+    "MS": "America/Chicago",        # Mississippi
+    "MT": "America/Denver",         # Montana
+    "NC": "America/New_York",       # North Carolina
+    "ND": "America/Chicago",        # North Dakota (split state)
+    "NE": "America/Chicago",        # Nebraska (split state)
+    "NH": "America/New_York",       # New Hampshire
+    "NJ": "America/New_York",       # New Jersey
+    "NM": "America/Denver",         # New Mexico
+    "NV": "America/Los_Angeles",    # Nevada
+    "NY": "America/New_York",       # New York
+    "OH": "America/New_York",       # Ohio
+    "OK": "America/Chicago",        # Oklahoma
+    "OR": "America/Los_Angeles",    # Oregon (split state)
+    "PA": "America/New_York",       # Pennsylvania
+    "PR": "America/Puerto_Rico",    # Puerto Rico; Not in CEMS
+    "RI": "America/New_York",       # Rhode Island
+    "SC": "America/New_York",       # South Carolina
+    "SD": "America/Chicago",        # South Dakota (split state)
+    "TN": "America/Chicago",        # Tennessee
+    "TX": "America/Chicago",        # Texas
+    "UT": "America/Denver",         # Utah
+    "VA": "America/New_York",       # Virginia
+    "VI": "America/Port_of_Spain",  # Virgin Islands; Not in CEMS
+    "VT": "America/New_York",       # Vermont
+    "WA": "America/Los_Angeles",    # Washington
+    "WI": "America/Chicago",        # Wisconsin
+    "WV": "America/New_York",       # West Virginia
+    "WY": "America/Denver",         # Wyoming
+    # Canada (none of these are in CEMS)
+    "AB": "America/Edmonton",       # Alberta
+    "BC": "America/Vancouver",      # British Columbia (split province)
+    "MB": "America/Winnipeg",       # Manitoba
+    "NB": "America/Moncton",        # New Brunswick
+    "NS": "America/Halifax",        # Nova Scotia
+    "NL": "America/St_Johns",       # Newfoundland and Labrador (split province)
+    "NT": "America/Yellowknife",    # Northwest Territories (split province)
+    "NU": "America/Iqaluit",        # Nunavut (split province)
+    "ON": "America/Toronto",        # Ontario (split province)
+    "PE": "America/Halifax",        # Prince Edwards Island
+    "QC": "America/Montreal",       # Quebec (split province)
+    "SK": "America/Regina",         # Saskatchewan  (split province)
+    "YT": "America/Whitehorse",     # Yukon Territory
+}
+"""
+Approximate mapping of US & Canadian jurisdictions to canonical timezones
+
+This is imperfect for states that have split timezones. See:
+https://en.wikipedia.org/wiki/List_of_time_offsets_by_U.S._state_and_territory
+For states that are split, the timezone that has more people in it.
+List of timezones in pytz.common_timezones
+Canada: https://en.wikipedia.org/wiki/Time_in_Canada#IANA_time_zone_database
+"""
+
+
+def find_timezone(*, lng=None, lat=None, state=None, strict=True):
+    """Find the timezone associated with the a specified input location.
+
+    Note that this function requires named arguments. The names are lng, lat,
+    and state.  lng and lat must be provided, but they may be NA. state isn't
+    required, and isn't used unless lng/lat are NA or timezonefinder can't find
+    a corresponding timezone.
+
+    Timezones based on states are imprecise, so it's far better to use lng/lat
+    if possible. If `strict` is True, state will not be used.
+    More on state-to-timezone conversion here:
+    https://en.wikipedia.org/wiki/List_of_time_offsets_by_U.S._state_and_territory
+
+    Args:
+        lng (int or float in [-180,180]): Longitude, in decimal degrees
+        lat (int or float in [-90, 90]): Latitude, in decimal degrees
+        state (str): Abbreviation for US state or Canadian province
+        strict (bool): Raise an error if no timezone is found?
+
+    Returns:
+        str: The timezone (as an IANA string) for that location.
+
+    Todo:
+        Update docstring.
+
+    """
+    try:
+        tz = TZ_FINDER.timezone_at(lng=lng, lat=lat)
+        if tz is None:  # Try harder
+            # Could change the search radius as well
+            tz = TZ_FINDER.closest_timezone_at(lng=lng, lat=lat)
+    # For some reason w/ Python 3.6 we get a ValueError here, but with
+    # Python 3.7 we get an OverflowError...
+    except (OverflowError, ValueError):
+        # If we're being strict, only use lng/lat, not state
+        if strict:
+            raise ValueError(
+                f"Can't find timezone for: lng={lng}, lat={lat}, state={state}"
+            )
+        tz = APPROXIMATE_TIMEZONES.get(state, None)
+    return tz
 
 
 def _occurrence_consistency(entity_id, compiled_df, col,
@@ -159,21 +289,21 @@ def _lat_long(dirty_df, clean_df, entity_id_df, entity_id,
     return ll_clean_df
 
 
-def _add_timezone(plants_entity):
-    """Adds plant IANA timezones from lat / lon.
+def _add_timezone(plants_entity: pd.DataFrame) -> pd.DataFrame:
+    """
+    Add plant IANA timezone based on lat/lon or state if lat/lon is unavailable.
 
     Args:
-        plants_entity (pandas.DataFrame): Plant entity table, including columns
-            named "latitude", "longitude", and optionally "state"
+        plants_entity: Plant entity table, including columns named "latitude",
+            "longitude", and optionally "state"
 
     Returns:
-        :class:`pandas.DataFrame`: A DataFrame containing the same table, with a
-        "timezone" column added. Timezone may be missing if lat / lon is
-        missing or invalid.
+        A DataFrame containing the same table, with a "timezone" column added.
+        Timezone may be missing if lat / lon is missing or invalid.
 
     """
     plants_entity["timezone"] = plants_entity.apply(
-        lambda row: pudl.helpers.find_timezone(
+        lambda row: find_timezone(
             lng=row["longitude"], lat=row["latitude"],
             state=row["state"], strict=False
         ),
